# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/orders/{id}/refunds": {
      "post": {
        "description": "Errors: NOTHING_TO_REFUND, REFUND_EXCEEDS_PAID and REFUND_EXCEEDS_ONLINE_PAID (409, details.refundablePaise), NO_ONLINE_PAYMENT (409), REFUND_GATEWAY_ERROR (502, with Razorpay's message; nothing is saved).",
        "operationId": "AdminRefundsController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refund money on an order, with a note the customer sees. Works in any status.",
        "tags": [
          "admin · orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Laundry API",
    "description": "Firebase phone-OTP ID token as Bearer. Money in paise. Dates are IST (YYYY-MM-DD).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "CUSTOMER",
          "DRIVER",
          "ADMIN",
          "SUPER_ADMIN"
        ]
      },
...
      "TimeSlot": {
        "type": "string",
        "enum": [
          "MORNING",
          "NOON",
          "EVENING"
        ]
      },
...
      "ItemUnit": {
        "type": "string",
        "enum": [
          "PIECE",
          "PAIR",
          "SET",
          "KG"
        ]
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "COD",
          "ONLINE"
        ]
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PICKUP_ASSIGNED",
          "PICKED_UP",
          "PROCESSING",
          "READY_FOR_DELIVERY",
          "DELIVERY_ASSIGNED",
          "OUT_FOR_DELIVERY",
          "DELIVERED",
          "CANCELLED"
        ]
      },
      "OrderSource": {
        "type": "string",
        "enum": [
          "CUSTOMER_APP",
          "ADMIN"
        ]
      },
      "OrderAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "formatted": {
            "type": "string"
          },
          "houseNo": {
            "type": "string"
          },
          "building": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string"
          },
          "area": {
            "type": "string",
            "nullable": true
          },
          "landmark": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pincode": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactPhone": {
            "type": "string"
          }
        },
        "required": [
          "label",
          "formatted",
          "houseNo",
          "building",
          "street",
          "area",
          "landmark",
          "city",
          "state",
          "pincode",
          "latitude",
          "longitude",
          "contactName",
          "contactPhone"
        ]
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "catalogItemId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemUnit"
              }
            ]
          },
          "unitPricePaise": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "lineTotalPaise": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "catalogItemId",
          "name",
          "unit",
          "unitPricePaise",
          "quantity",
          "lineTotalPaise"
        ]
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "UNPAID",
          "PARTIALLY_PAID",
          "PAID",
          "REFUNDED"
        ]
      },
      "PersonRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone"
        ]
      },
      "OrderEventDto": {
        "type": "object",
        "properties": {
          "fromStatus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "toStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "actorRole": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "fromStatus",
          "toStatus",
          "actorRole",
          "note",
          "createdAt"
        ]
      },
      "RefundMethod": {
        "type": "string",
        "enum": [
          "RAZORPAY",
          "CASH",
          "BANK_TRANSFER"
        ]
      },
      "RefundStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSED",
          "FAILED"
        ],
        "description": "PENDING: Razorpay is still sending it. Only PENDING and PROCESSED count toward refundedPaise."
      },
      "StaffRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "RefundDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "amountPaise": {
            "type": "number"
          },
          "method": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RefundMethod"
              }
            ]
          },
          "status": {
            "description": "PENDING: Razorpay is still sending it. Only PENDING and PROCESSED count toward refundedPaise.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RefundStatus"
              }
            ]
          },
          "note": {
            "type": "string",
            "description": "Staff note to the customer."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "processedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "failureReason": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "nullable": true,
            "description": "Staff member who issued it. Admin view only.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StaffRefDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "amountPaise",
          "method",
          "status",
          "note",
          "createdAt",
          "processedAt",
          "failureReason"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string",
            "example": "LD001042"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderSource"
              }
            ]
          },
          "pickupDate": {
            "type": "string"
          },
          "pickupSlot": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSlot"
              }
            ]
          },
          "pickupSlotLabel": {
            "type": "string"
          },
          "deliveryDate": {
            "type": "string"
          },
          "deliverySlot": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSlot"
              }
            ]
          },
          "deliverySlotLabel": {
            "type": "string"
          },
          "pickupAddress": {
            "$ref": "#/components/schemas/OrderAddressDto"
          },
          "deliveryAddress": {
            "$ref": "#/components/schemas/OrderAddressDto"
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "subtotalPaise": {
            "type": "number"
          },
          "discountPaise": {
            "type": "number"
          },
          "deliveryFeePaise": {
            "type": "number"
          },
          "totalPaise": {
            "type": "number"
          },
          "paidPaise": {
            "type": "number"
          },
          "refundedPaise": {
            "type": "number",
            "description": "Sum of PENDING and PROCESSED refunds."
          },
          "refundablePaise": {
            "type": "number",
            "description": "What can still be refunded (paidPaise - refundedPaise). Admin view only."
          },
          "amountDuePaise": {
            "type": "number"
          },
          "promoCode": {
            "type": "string",
            "nullable": true
          },
          "paymentMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              }
            ]
          },
          "paymentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "customer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonRefDto"
              }
            ]
          },
          "pickupDriver": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonRefDto"
              }
            ]
          },
          "deliveryDriver": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonRefDto"
              }
            ]
          },
          "allowedNextStatuses": {
            "type": "array",
            "description": "Statuses the caller is allowed to move this order to.",
            "items": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          "dispatchFailedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelReason": {
            "type": "string",
            "nullable": true
          },
          "pickedUpAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deliveredAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderEventDto"
            }
          },
          "refunds": {
            "description": "Oldest first. Customer and admin views only.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefundDto"
            }
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "source",
          "pickupDate",
          "pickupSlot",
          "pickupSlotLabel",
          "deliveryDate",
          "deliverySlot",
          "deliverySlotLabel",
          "pickupAddress",
          "deliveryAddress",
          "instructions",
          "items",
          "subtotalPaise",
          "discountPaise",
          "deliveryFeePaise",
          "totalPaise",
          "paidPaise",
          "refundedPaise",
          "amountDuePaise",
          "promoCode",
          "paymentMethod",
          "paymentStatus",
          "customer",
          "pickupDriver",
          "deliveryDriver",
          "allowedNextStatuses",
          "dispatchFailedAt",
          "cancelReason",
          "pickedUpAt",
          "deliveredAt",
          "cancelledAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateRefundDto": {
        "type": "object",
        "properties": {
          "amountPaise": {
            "type": "number",
            "description": "Amount to refund, in paise. At most what was paid minus earlier refunds.",
            "minimum": 1
          },
          "method": {
            "description": "RAZORPAY sends the money back to the card/UPI the customer paid with. CASH and BANK_TRANSFER record money staff already returned outside the app.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RefundMethod"
              }
            ]
          },
          "note": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500,
            "description": "Shown to the customer in the app and in their notification"
          }
        },
        "required": [
          "amountPaise",
          "method",
          "note"
        ]
      },
...
    }
  }
}